Moved statistics to own file so homepage app will not load the statistics view which also loads rest framework, causing a circular error
